Fix night sync.
toggleNight is used not only by the user to switch between day and
night, but also by the socket handling code to push values from the host
to the clients. In that case, the client must accept the value the host
has declared.

This mirrors the behavior of the pre-existing `toggle`, which was
replaced in 66a36d74225d164657da2fa94861ee0feea8112d.

h/t to Wedge for identifying the culprit commit.