Add search function to the role modal.
When assigning a role to a player, it is now possible to type a case-
and whitespace-insensitive prefix of it to quickly find the desired
role. Pressing Enter will set the role, if the query is specific enough
to have a unique match.

I've found this very convenient to quickly fill in a grimoire without
having to hunt for a specific token.